+++ src/app/components/loading-spinner/service/terra-loading-spinner.service.ts
@@ -15,7 +15,7 @@ export class TerraLoadingSpinnerService
 {
     public observable:Observable<boolean>;
 
-    private isLoading:boolean = false;
+    private _isLoading:boolean = false;
     private _subscriber:Subscriber<boolean>;
 
     constructor(private zone:NgZone)
@@ -29,19 +29,19 @@ export class TerraLoadingSpinnerService
     public start():void
     {
         // check if currently not loading to reduce unnecessary change detections
-        if(this._subscriber && !this.isLoading)
+        if(this._subscriber && !this._isLoading)
         {
-            this.isLoading = true;
-            this._subscriber.next(this.isLoading);
+            this._isLoading = true;
+            this._subscriber.next(this._isLoading);
         }
     }
 
     public stop():void
     {
         // check if currently loading to reduce unnecessary change detections
-        if(this._subscriber && this.isLoading)
+        if(this._subscriber && this._isLoading)
         {
-            this.isLoading = false;
+            this._isLoading = false;
 
             // to send no change detection run the setTimeout outside of angular
             this.zone.runOutsideAngular(() =>
@@ -49,17 +49,22 @@ export class TerraLoadingSpinnerService
                     // set timeout to stop the loading-spinner from blinking because of sequential started events
                     setTimeout(() =>
                     {
-                        if(!this.isLoading)
+                        if(!this._isLoading)
                         {
                             // run inside angular zone to detect changes from isLoading to false
                             this.zone.run(() =>
                             {
-                                this._subscriber.next(this.isLoading);
+                                this._subscriber.next(this._isLoading);
                             });
                         }
                     }, 100);
                 }
             );
         }
+    }
+
+    public get isLoading():boolean
+    {
+        return this._isLoading;
     }
 }