+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+<a name="2.3.3"></a>
+# 2.3.3 (20.07.2018)
+
+### Bug Fixes
+* **terra-dynamic-form** added possibility to change the debounce time
+
+### Features
+* **terra-loading-spinner** added public getter for `isLoading` property, since this information is a global one and can be used in any component to disable elements while a request is pending.
+* **terra-base-service** added public getter for `isLoading` property to be available in every specific service extension.
+
 <a name="2.3.2"></a>
 # 2.3.2 (20.07.2018)